+++ tests/pallas/gpu_paged_attention_test.py
@@ -14,11 +14,18 @@
 
 import sys
 import warnings
+
 from absl.testing import absltest
 import jax
 from jax._src import test_util as jtu
 import jax.numpy as jnp
 import numpy as np
+
+warnings.filterwarnings(
+    "ignore",
+    message="jax.experimental.pallas.ops.gpu.* is deprecated.*",
+    category=DeprecationWarning,
+)
 
 if sys.platform != "win32":
   from jax.experimental.pallas.ops.gpu import paged_attention